cli/linker: add wide address parsing for go/bin/output spans Accept 4-8 hex digit addresses for --go and --bin ranges. Promote .output image/loadaddr metadata to u32. Support wide binary ranges and HEX start-linear-address record emission. Keep PRG load address constrained to 16-bit with explicit diagnostics. Add regression tests for wide CLI parsing, wide output metadata, and wide HEX/bin behavior.

+++ src/assembler/bootstrap.rs
@@ -211,10 +211,10 @@ pub(crate) struct LinkerOutputDirective {
     pub(crate) format: LinkerOutputFormat,
     pub(crate) sections: Vec<String>,
     pub(crate) contiguous: bool,
-    pub(crate) image_start: Option<u16>,
-    pub(crate) image_end: Option<u16>,
+    pub(crate) image_start: Option<u32>,
+    pub(crate) image_end: Option<u32>,
     pub(crate) fill: Option<u8>,
-    pub(crate) loadaddr: Option<u16>,
+    pub(crate) loadaddr: Option<u32>,
 }
 
 impl RootMetadata {
